requirements.txt
fastapi[all]
transformers
torch
pillow
